remove menu option for booking marking avatar entities
