+++ interface/src/AvatarBookmarks.cpp
@@ -19,6 +19,11 @@
 #include <Application.h>
 #include <OffscreenUi.h>
 #include <avatar/AvatarManager.h>
+#include <EntityItemID.h>
+#include <EntityTree.h>
+#include <PhysicalEntitySimulation.h>
+#include <EntityEditPacketSender.h>
+#include <VariantMapToScriptValue.h>
 
 #include "MainWindow.h"
 #include "Menu.h"
@@ -29,6 +34,62 @@
 
 #include <QtQuick/QQuickWindow>
 
+
+void addAvatarEntities(const QVariantList& avatarEntities) {
+    auto nodeList = DependencyManager::get<NodeList>();
+    const QUuid myNodeID = nodeList->getSessionUUID();
+    EntityTreePointer entityTree = DependencyManager::get<EntityTreeRenderer>()->getTree();
+    if (!entityTree) {
+        return;
+    }
+    EntitySimulationPointer entitySimulation = entityTree->getSimulation();
+    PhysicalEntitySimulationPointer physicalEntitySimulation = std::static_pointer_cast<PhysicalEntitySimulation>(entitySimulation);
+    EntityEditPacketSender* entityPacketSender = physicalEntitySimulation->getPacketSender();
+    QScriptEngine scriptEngine;
+    for (int index = 0; index < avatarEntities.count(); index++) {
+        const QVariantMap& avatarEntityProperties = avatarEntities.at(index).toMap();
+        QVariant variantProperties = avatarEntityProperties["properties"];
+        QVariantMap asMap = variantProperties.toMap();
+        QScriptValue scriptProperties = variantMapToScriptValue(asMap, scriptEngine);
+        EntityItemProperties entityProperties;
+        EntityItemPropertiesFromScriptValueHonorReadOnly(scriptProperties, entityProperties);
+
+        entityProperties.setParentID(myNodeID);
+        entityProperties.setClientOnly(true);
+        entityProperties.setOwningAvatarID(myNodeID);
+        entityProperties.setSimulationOwner(myNodeID, AVATAR_ENTITY_SIMULATION_PRIORITY);
+        entityProperties.markAllChanged();
+
+        EntityItemID id = EntityItemID(QUuid::createUuid());
+        bool success = true;
+        entityTree->withWriteLock([&] {
+            EntityItemPointer entity = entityTree->addEntity(id, entityProperties);
+            if (entity) {
+                if (entityProperties.queryAACubeRelatedPropertyChanged()) {
+                    // due to parenting, the server may not know where something is in world-space, so include the bounding cube.
+                    bool success;
+                    AACube queryAACube = entity->getQueryAACube(success);
+                    if (success) {
+                        entityProperties.setQueryAACube(queryAACube);
+                    }
+                }
+
+                entity->setLastBroadcast(usecTimestampNow());
+                // since we're creating this object we will immediately volunteer to own its simulation
+                entity->flagForOwnershipBid(VOLUNTEER_SIMULATION_PRIORITY);
+                entityProperties.setLastEdited(entity->getLastEdited());
+            } else {
+                qCDebug(entities) << "AvatarEntitiesBookmark failed to add new Entity to local Octree";
+                success = false;
+            }
+        });
+
+        if (success) {
+            entityPacketSender->queueEditEntityMessage(PacketType::EntityAdd, entityTree, id, entityProperties);
+        }
+    }
+}
+
 AvatarBookmarks::AvatarBookmarks() {
     _bookmarksFilename = PathUtils::getAppDataPath() + "/" + AVATARBOOKMARKS_FILENAME;
     readFromFile();
@@ -38,7 +99,7 @@ void AvatarBookmarks::readFromFile() {
     // migrate old avatarbookmarks.json, used to be in 'local' folder on windows
     QString oldConfigPath = QStandardPaths::writableLocation(QStandardPaths::DataLocation) + "/" + AVATARBOOKMARKS_FILENAME;
     QFile oldConfig(oldConfigPath);
-    
+
     // I imagine that in a year from now, this code for migrating (as well as the two lines above)
     // may be removed since all bookmarks should have been migrated by then
     // - Robbie Uvanni (6.8.2017)
@@ -48,9 +109,9 @@ void AvatarBookmarks::readFromFile() {
         } else {
             qCDebug(interfaceapp) << "Failed to migrate" << AVATARBOOKMARKS_FILENAME;
         }
-    } 
-    
-    Bookmarks::readFromFile();     
+    }
+
+    Bookmarks::readFromFile();
 }
 
 void AvatarBookmarks::setupMenus(Menu* menubar, MenuWrapper* menu) {
@@ -81,23 +142,27 @@ void AvatarBookmarks::changeToBookmarkedAvatar() {
         myAvatar->useFullAvatarURL(action->data().toString());
         qCDebug(interfaceapp) << " Using Legacy V1 Avatar Bookmark ";
     } else {
-        
+
         const QMap<QString, QVariant> bookmark = action->data().toMap();
-        // Not magic value. This is the current made version, and if it changes this interpreter should be updated to 
+        // Not magic value. This is the current made version, and if it changes this interpreter should be updated to
         // handle the new one separately.
         // This is where the avatar bookmark entry is parsed. If adding new Value, make sure to have backward compatability with previous
         if (bookmark.value(ENTRY_VERSION) == 3) {
-                const QString& avatarUrl = bookmark.value(ENTRY_AVATAR_URL, "").toString();
-                myAvatar->useFullAvatarURL(avatarUrl);
-                qCDebug(interfaceapp) << "Avatar On " << avatarUrl;
-                const QList<QVariant>& attachments = bookmark.value(ENTRY_AVATAR_ATTACHMENTS, QList<QVariant>()).toList();
+            myAvatar->removeAvatarEntities();
+            const QString& avatarUrl = bookmark.value(ENTRY_AVATAR_URL, "").toString();
+            myAvatar->useFullAvatarURL(avatarUrl);
+            qCDebug(interfaceapp) << "Avatar On " << avatarUrl;
+            const QList<QVariant>& attachments = bookmark.value(ENTRY_AVATAR_ATTACHMENTS, QList<QVariant>()).toList();
 
-                qCDebug(interfaceapp) << "Attach " << attachments;
-                myAvatar->setAttachmentsVariant(attachments);
+            qCDebug(interfaceapp) << "Attach " << attachments;
+            myAvatar->setAttachmentsVariant(attachments);
 
-                const float& qScale = bookmark.value(ENTRY_AVATAR_SCALE, 1.0f).toFloat();
-                myAvatar->setAvatarScale(qScale);
-     
+            const float& qScale = bookmark.value(ENTRY_AVATAR_SCALE, 1.0f).toFloat();
+            myAvatar->setAvatarScale(qScale);
+
+            const QVariantList& avatarEntities = bookmark.value(ENTRY_AVATAR_ENTITIES, QVariantList()).toList();
+            addAvatarEntities(avatarEntities);
+
         } else {
             qCDebug(interfaceapp) << " Bookmark entry does not match client version, make sure client has a handler for the new AvatarBookmark";
         }
@@ -126,6 +191,7 @@ void AvatarBookmarks::addBookmark() {
         bookmark.insert(ENTRY_AVATAR_URL, avatarUrl);
         bookmark.insert(ENTRY_AVATAR_SCALE, avatarScale);
         bookmark.insert(ENTRY_AVATAR_ATTACHMENTS, myAvatar->getAttachmentsVariant());
+        bookmark.insert(ENTRY_AVATAR_ENTITIES, myAvatar->getAvatarEntitiesVariant());
 
         Bookmarks::addBookmarkToFile(bookmarkName, bookmark);
     });
